Extract the handoff and seal routes out of submissions.ts
`POST /api/submissions/:token/handoff` and `POST /api/submissions/:token/seal`
are one cluster — both are the creator changing course on a round that already
exists, both re-enter the build lane, and neither belongs to the status
derivation that fills most of submissions.ts. They move to
`creation/handoff-seal-routes.ts` behind `registerHandoffSealRoutes`, following
the registrar-with-deps shape the other eleven extractions use.

The route bodies are unchanged. Ten registrar-private closures cross as deps
(`store`, `gamesStore`, `githubClient`, `submissionTokenSecret`,
`managedAvailabilityGate`, `now`, `checkUserAccess`, `builderOf`,
`invalidateStatusCache`, `resumeBuild`) plus the gate trigger. The two
`options.agentChannel?.…` reads the seal handler did per request are now read
once at registration and passed in — the same object either way, since
`agentChannel` is fixed before the registrar runs. `ResumeOutcome` is declared
structurally rather than imported, because submissions.ts imports this file.

Proved by diffing both route bodies with comments and blank lines stripped:
identical apart from those two `const` lines. The prose comments were rewritten
as `//` one-liners so the new file starts at zero comment-prose debt.

submissions.ts: 3489 → 3171 lines.

Also fixes a race in the token-info test added last batch: it asserted
`expiresInDays === 89`, which is only true once a millisecond has elapsed since
the mint. It now accepts either 89 or 90, and the flooring it meant to pin is
tested separately by minting off-clock, where the half-day remainder is real.
