+++ apps/api/src/platform/access-token-routes.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { generateAccessToken } from './access-token.js';
-import { DEFAULT_EXPIRY_DAYS, MAX_TOKENS_PER_UID } from './access-token-service.js';
+import { DEFAULT_EXPIRY_DAYS, MAX_TOKENS_PER_UID, mintAccessTokenFor } from './access-token-service.js';
 import { buildApp } from './app.js';
 import { mintSessionToken, SESSION_COOKIE_NAME } from './auth.js';
 import { InMemoryStore } from './store.js';
@@ -14,6 +14,7 @@ import { InMemoryStore } from './store.js';
  */
 
 const sessionSecret = 'dev-session-secret-change-me';
+const HOUR = 60 * 60 * 1000;
 
 function sessionHeaders(uid: string) {
   return { cookie: `${SESSION_COOKIE_NAME}=${mintSessionToken(uid, sessionSecret)}` };
@@ -297,9 +298,30 @@ describe('authenticating with a personal access token', () => {
     const res = await app.inject({ method: 'GET', url: '/api/auth/token-info', headers: bearer(token) });
     expect(res.statusCode).toBe(200);
     expect(res.json().name).toBe('github actions');
-    expect(res.json().expiresInDays).toBe(DEFAULT_EXPIRY_DAYS - 1);
+    // A whole day may or may not have been consumed by the time this reads back.
+    expect(res.json().expiresInDays).toBeGreaterThanOrEqual(DEFAULT_EXPIRY_DAYS - 1);
+    expect(res.json().expiresInDays).toBeLessThanOrEqual(DEFAULT_EXPIRY_DAYS);
     // The date has to be readable, not just the day count.
     expect(Date.parse(res.json().expiresAt)).toBeGreaterThan(Date.now());
+  });
+
+  // Minted off-clock so the half-day remainder is real, not a race.
+  it.each([
+    [12 * HOUR, 7],
+    [-12 * HOUR, 6],
+  ])('floors the day count rather than rounding it (offset %i)', async (offsetMs, expected) => {
+    const app = await appWith(store);
+    const { token } = await mintAccessTokenFor(store, {
+      uid: 'bot:ci',
+      name: 'clock probe',
+      createdByUid: 'g:boss',
+      expiresInDays: 7,
+      nowMs: Date.now() + offsetMs,
+    });
+
+    const res = await app.inject({ method: 'GET', url: '/api/auth/token-info', headers: bearer(token) });
+    expect(res.statusCode).toBe(200);
+    expect(res.json().expiresInDays).toBe(expected);
   });
 
   it('never leaks credential material through token-info', async () => {
